+++ packages/forms/src/components/FloatingSearch/messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from 'react-intl';
+
+const messages = defineMessages({
+    clearButton: {
+        id: 'FloatingSearch.clearButton.text',
+        defaultMessage: 'Clear',
+        description: 'Text of the button to clear the FloatingSearch',
+    },
+});
+
+export default messages;