+++ packages/forms/src/index.js
@@ -1,2 +1,4 @@
 export { default as UniversalPicker } from './components/UniversalPicker';
 export { default as UniversalPickerOption } from './components/UniversalPickerOption';
+export { default as FloatingSearch } from './components/FloatingSearch';
+export { default as FloatingSearchButtonIcon } from './components/FloatingSearchButtonIcon';