+++ packages/listview/src/components/BatchActionsBar/styled.js
@@ -5,11 +5,11 @@ import { BORDER_RADIUS_2 } from 'react-rainbow-components/styles/borderRadius';
 import { FONT_SIZE_TEXT_LARGE } from 'react-rainbow-components/styles/fontSizes';
 
 export const StyledContainer = attachThemeAttrs(styled.div)`
-    padding: 16px;
+    padding: 8px 8px 8px 12px;
     background-color: ${(props) => props.palette.background.main};
     color: ${(props) => props.palette.text.main};
     border: 1px solid transparent;
-    box-shadow: 0 10px 28px 0 rgba(0, 0, 0, 0.2);
+    box-shadow: 0 4px 30px 0 rgba(0, 0, 0, 0.3);
     border-radius: ${BORDER_RADIUS_2};
     transition: all 0.1s linear, padding 0s, border 0s;
     line-height: 2.5rem;
@@ -33,12 +33,12 @@ export const StyledRightContent = styled.div`
 
 export const StyledBadge = styled(Badge)`
     flex: 0 0 auto;
-    font-size: 20px;
+    font-size: 18px;
     margin-right: 16px;
-    height: 40px;
-    line-height: 40px;
+    height: 36px;
+    line-height: 36px;
     font-weight: 800;
-    padding: 0 16px;
+    padding: 0 12px;
 `;
 
 export const StyledLabel = attachThemeAttrs(styled.span)`