+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.12.0](https://github.com/nexxtway/rainbow-modules/compare/v0.11.0...v0.12.0) (2020-11-17)
+
+**Note:** Version bump only for package @rainbow-modules/icons
+
+# [0.11.0](https://github.com/nexxtway/rainbow-modules/compare/v0.10.0...v0.11.0) (2020-11-13)
+
+### Features
+
+-   added camera to icons module ([#193](https://github.com/nexxtway/rainbow-modules/issues/193)) ([9ac529f](https://github.com/nexxtway/rainbow-modules/commit/9ac529f76ed09591177ce7327147c05f1fec327c))
+
 # [0.10.0](https://github.com/nexxtway/rainbow-modules/compare/v0.8.0...v0.10.0) (2020-10-29)
 
 ### Bug Fixes